Orville.PostgreSQL.AutoMigration

Documentation

autoMigrateSchema :: MonadOrville m => [SchemaItem] -> m ()

This function compares the list of SchemaItems provided against the current schema found in the database to determine whether any migration are necessary. If any changes need to be made, this function executes. You can call generateMigrationSteps and executeMigrationSteps yourself if you want to have more control over the process.

generateMigrationSteps :: MonadOrville m => [SchemaItem] -> m [MigrationStep]

Compares the list of SchemaItems provided against the current schema found in the database and returns a list of MigrationSteps that could be executed to make the database schema match the items given.

You can execute the MigrationSteps yourself using executeVoid, or use the executeMigrationSteps convenience function.

executeMigrationSteps :: MonadOrville m => [MigrationStep] -> m ()

A convenience function for executing a list of MigrationSteps that has be previously devised via generateMigrationSteps.

data SchemaItem where

A SchemaItem represents a single item in a database schema such as a table, index or constraint. The constructor functions below can be used to create items from other types (such as TableDefinition) to put them into a list to be used with autoMigrateSchema.

Constructors

SchemaTable :: TableDefinition key writeEntity readEntity -> SchemaItem
Constructs a SchemaItem from a TableDefinition.

SchemaDropTable :: TableIdentifier -> SchemaItem
Constructs a SchemaItem that will drop the specified table if it is found in the database.

SchemaSequence :: SequenceDefinition -> SchemaItem
Constructs a SchemaItem from a SequenceDefinition.

SchemaDropSequence :: SequenceIdentifier -> SchemaItem
Constructs a SchemaItem that will drop the specified table if it is found in the database.

schemaItemSummary :: SchemaItem -> String

Returns a one-line string describe the SchemaItem, suitable for a human to identify it in a list of output.

For example, a SchemaItem constructed via schemaTable gives Table <table name>.

data MigrationStep

A single SQL statement that will be executed in order to migrate the database to the desired result. You can use generateMigrationSteps to get a list of these yourself for inspection and debugging.

data MigrationDataError

A MigrationDataError will be thrown from the migration functions if data necessary for migration cannot be found.