Orville.PostgreSQL.Connection

Description

Documentation

data Connection

The basic connection interface.

data Pool a

data ConnectionUsedAfterCloseError

data ConnectionError

data SqlExecutionError

Constructors

SqlExecutionError

Fields

sqlExecutionErrorExecStatus :: ExecStatus

sqlExecutionErrorMessage :: ByteString

sqlExecutionErrorSqlState :: Maybe ByteString

sqlExecutionErrorSqlQuery :: ByteString

data NoticeReporting

An option for createConnectionPool than indicates whether the LibPQ should print notice reports for warnings to the console

Constructors

EnableNoticeReporting

DisableNoticeReporting

createConnectionPool

Arguments

:: NoticeReporting
Whether or not notice reporting from LibPQ should be enabled

-> Int
Number of stripes in the connection pool

-> NominalDiffTime
Linger time before closing an idle connection

-> Int
Max number of connections to allocate per stripe

-> ByteString
A PostgreSQL connection string

-> IO (Pool Connection)

createConnectionPool allocates a pool of connections to a PosgreSQL server.

executeRaw :: Connection -> ByteString -> [Maybe PgTextFormatValue] -> IO Result

executeRaw runs a given SQL statement returning the raw underlying result.

All handling of stepping through the result set is left to the caller. This potentially leaves connections open much longer than one would expect if all of the results are not iterated through immediately *and* the data copied. Use with caution.

escapeStringLiteral :: Connection -> ByteString -> IO ByteString

Escapes a string for use as a literal within a SQL command that will be execute on the given connection. This uses the PQescapeStringConn function from libpq, which takes the character encoding of the connection into account. This function only escapes the characters to be used in as a string literal -- it does not add the surrounding quotes.