Orville.PostgreSQL.EntityTrace

Documentation

type EntityTraceT trace = ReaderT (EntityTraceState trace)

EntityTraceT is simply a synonym for ReaderT over an EntityTraceState. The EntityTraceState keeps track of what traces have been recorded, including keeping them in sync with any transaction or savepoint operations that are performed via the withTransaction function. I.E. if a transaction or savepoint is rolled-back any traces that had been recorded by insertEntityTraced, updateEntityTrace, or deletedEntityTraced are also rolled-back.

data EntityTraceState trace

An EntityTraceState is created by runEntityTraceT to keep track of the state of traces that are to be returned when the function finishes.

runEntityTraceT :: MonadIO n => ErrorDetailLevel -> Pool Connection -> (OrvilleState -> m a -> n a) -> EntityTraceT trace m a -> n (a, [trace])

runEntityTraceT runs an Orville action that has tracing behavior and returns the traces that were committed. Note that there will never be any uncommitted traces at the end because any withTransaction block must by contained within the action passed to runEntityTraceT. This function creates a new OrvilleState and passes it to the run function that you provide in order to enforce this.

Note that if an exception occurs in m and is not caught within the action passed to runEntityTraceT, you will lose any traces that may have happened up to the point of the action, including those related to database operations that were successfully committed. If you wish to respond to those traces, you need to catch the exception inside the action given to runEntityTraceT so that it can return the traces to you.

insertEntityTraced :: MonadEntityTrace trace m => TracedTable trace key writeEntity readEntity -> writeEntity -> m readEntity

Inserts an entity into the database via insertAndReturnEntity, then records any traces returned by the addInsertTrace callbacks on the TracedTable via recordTraces.

See runEntityTraceT for a stock way to run actions using these trace operations and get access to the traces that are recorded.

updateEntityTraced :: MonadEntityTrace trace m => TracedTable trace key writeEntity readEntity -> readEntity -> writeEntity -> m ()

Updates an entity in the database via updateAndReturnEntity, then records any traces returned by the addUpdateTrace callbacks on the TracedTable via recordTraces.

See runEntityTraceT for a stock way to run actions using these trace operations and get access to the traces that are recorded.

deleteEntityTraced :: MonadEntityTrace trace m => TracedTable trace key writeEntity readEntity -> key -> m ()

Updates an entity in the database via deleteAndReturnEntity, then records any traces returned by the addDeleteTrace callbacks on the TracedTable via recordTraces.

See runEntityTraceT for a stock way to run actions using these trace operations and get access to the traces that are recorded.

data TracedTable trace key writeEntity readEntity

A TracedTable is a regular TableDefinition with callbacks for constructing the trace items that will be captured when functions such as insertEntityTraced are used. Use mkTracedTable to build a TracedTable with no callbacks and then use addInsertTrace and friends to register callbacks callbacks on the table.

mkTracedTable :: (readEntity -> key) -> TableDefinition (HasKey key) writeEntity readEntity -> TracedTable trace key writeEntity readEntity

Constructs a TracedTable that wraps the given TableDefinition so that it can be used with functions such as insertEntityTraced. The resulting TracedTable has no trace callbacks initially. You can addInsertTrace and similar functions ta add callbacks.

addInsertTrace :: (writeEntity -> readEntity -> [trace]) -> TracedTable trace key writeEntity readEntity -> TracedTable trace key writeEntity readEntity

Adds a callback that will be used during insertEntityTraced to build the trace items to be recorded for that insert. The trace building function will be passed the writeEntity given to insertEntityTraced as well as the readEntity that is returned as a result of the insert.

addUpdateTrace :: (readEntity -> writeEntity -> Maybe readEntity -> [trace]) -> TracedTable trace key writeEntity readEntity -> TracedTable trace key writeEntity readEntity

Adds a callback that will be used during updateEntityTraced to build the trace items to be recorded for that update. The trace building function will be passed the readEntity and writeEntity that were given to updateEntityTraced as well as the any readEntity that is returned as a result of the insert. If the update matches no rows in the database, Nothing will be passed as the last argument to the trace building function.

addDeleteTrace :: (key -> Maybe readEntity -> [trace]) -> TracedTable trace key writeEntity readEntity -> TracedTable trace key writeEntity readEntity

Adds a callback that will be used during deleteEntityTraced to build the trace items to be recorded for that update. The trace building function will be passed the key that sa given to deleteEntityTraced as well as the any readEntity that is returned as a result of the delete. If the delete matches no rows in the database, Nothing will be passed as the last argument to the trace building function.

untracedTableDefinition :: TracedTable trace key writeEntity readEntity -> TableDefinition (HasKey key) writeEntity readEntity

Returns the TableDefinition underlying a TracedTable. You can use this to get access to the table definiton to use the regular untraced Orville functions without having to keep a separate reference to the untraced table around.

class Monad m => MonadEntityTrace trace m | m -> trace where

MonadEntityTrace provides the interface used by functions such as insertEntityTraced to perform tracing as part of their operation. An instance for 'ReaderT EntityTraceState' is provided which can be used via runEntityTraceT.

Methods

recordTraces :: [trace] -> m ()

liftOrvilleUntraced :: (forall n. MonadOrville n => n a) -> m a