Orville.PostgreSQL.Internal.Expr

Description

Documentation

data WhereClause

whereClause :: BooleanExpr -> WhereClause

data ComparisonOperator

equalsOp :: ComparisonOperator

notEqualsOp :: ComparisonOperator

greaterThanOp :: ComparisonOperator

lessThanOp :: ComparisonOperator

greaterThanOrEqualsOp :: ComparisonOperator

lessThanOrEqualsOp :: ComparisonOperator

likeOp :: ComparisonOperator

iLikeOp :: ComparisonOperator

data BooleanExpr

orExpr :: BooleanExpr -> BooleanExpr -> BooleanExpr

andExpr :: BooleanExpr -> BooleanExpr -> BooleanExpr

parenthesized :: BooleanExpr -> BooleanExpr

columnEquals :: ColumnName -> SqlValue -> BooleanExpr

columnNotEquals :: ColumnName -> SqlValue -> BooleanExpr

columnGreaterThan :: ColumnName -> SqlValue -> BooleanExpr

columnLessThan :: ColumnName -> SqlValue -> BooleanExpr

columnGreaterThanOrEqualTo :: ColumnName -> SqlValue -> BooleanExpr

columnLessThanOrEqualTo :: ColumnName -> SqlValue -> BooleanExpr

columnLike :: ColumnName -> SqlValue -> BooleanExpr

columnLikeInsensitive :: ColumnName -> SqlValue -> BooleanExpr

columnIsNull :: ColumnName -> BooleanExpr

columnIsNotNull :: ColumnName -> BooleanExpr

comparison :: ValueExpression -> ComparisonOperator -> ValueExpression -> BooleanExpr

columnIn :: ColumnName -> NonEmpty SqlValue -> BooleanExpr

columnNotIn :: ColumnName -> NonEmpty SqlValue -> BooleanExpr

columnTupleIn :: NonEmpty ColumnName -> NonEmpty (NonEmpty SqlValue) -> BooleanExpr

Checks that the tuple constructed from the given columns in one of the tuples specified in the input list. It is up to the caller to ensure that all the tuples given have the same arity.

columnTupleNotIn :: NonEmpty ColumnName -> NonEmpty (NonEmpty SqlValue) -> BooleanExpr

Checks that the tuple constructed from the given columns is NOT one of the tuples specified in the input list. It is up to the caller to ensure that all the tuples given have the same arity.

data InValuePredicate

inPredicate :: ValueExpression -> InValuePredicate -> BooleanExpr

notInPredicate :: ValueExpression -> InValuePredicate -> BooleanExpr

inValueList :: NonEmpty ValueExpression -> InValuePredicate

data ValueExpression

columnReference :: ColumnName -> ValueExpression

valueExpression :: SqlValue -> ValueExpression

rowValueConstructor :: NonEmpty ValueExpression -> ValueExpression

functionCall :: FunctionName -> [ValueExpression] -> ValueExpression

data SetClauseList

setClauseList :: NonEmpty SetClause -> SetClauseList

data SetClause

setColumn :: ColumnName -> SqlValue -> SetClause

data UpdateExpr

updateExpr :: Qualified TableName -> SetClauseList -> Maybe WhereClause -> Maybe ReturningExpr -> UpdateExpr

data BeginTransactionExpr

beginTransaction :: Maybe TransactionMode -> BeginTransactionExpr

data TransactionMode

readWrite :: TransactionMode

readOnly :: TransactionMode

deferrable :: TransactionMode

notDeferrable :: TransactionMode

isolationLevel :: IsolationLevel -> TransactionMode

data IsolationLevel

serializable :: IsolationLevel

repeatableRead :: IsolationLevel

readCommitted :: IsolationLevel

readUncommitted :: IsolationLevel

data CommitExpr

commit :: CommitExpr

data RollbackExpr

rollback :: RollbackExpr

rollbackTo :: SavepointName -> RollbackExpr

data SavepointExpr

savepoint :: SavepointName -> SavepointExpr

data ReleaseSavepointExpr

releaseSavepoint :: SavepointName -> ReleaseSavepointExpr

data CreateTableExpr

createTableExpr :: Qualified TableName -> [ColumnDefinition] -> Maybe PrimaryKeyExpr -> [TableConstraint] -> CreateTableExpr

data PrimaryKeyExpr

primaryKeyExpr :: NonEmpty ColumnName -> PrimaryKeyExpr

data AlterTableExpr

alterTableExpr :: Qualified TableName -> NonEmpty AlterTableAction -> AlterTableExpr

data AlterTableAction

addColumn :: ColumnDefinition -> AlterTableAction

dropColumn :: ColumnName -> AlterTableAction

addConstraint :: TableConstraint -> AlterTableAction

dropConstraint :: ConstraintName -> AlterTableAction

alterColumnType :: ColumnName -> DataType -> Maybe UsingClause -> AlterTableAction

alterColumnSetDefault :: SqlExpression valueExpression => ColumnName -> valueExpression -> AlterTableAction

alterColumnDropDefault :: ColumnName -> AlterTableAction

data UsingClause

usingCast :: ColumnName -> DataType -> UsingClause

alterColumnNullability :: ColumnName -> AlterNotNull -> AlterTableAction

data AlterNotNull

setNotNull :: AlterNotNull

dropNotNull :: AlterNotNull

data DropTableExpr

dropTableExpr :: Maybe IfExists -> Qualified TableName -> DropTableExpr

data TableConstraint

uniqueConstraint :: NonEmpty ColumnName -> TableConstraint

foreignKeyConstraint :: NonEmpty ColumnName -> Qualified TableName -> NonEmpty ColumnName -> Maybe ForeignKeyUpdateActionExpr -> Maybe ForeignKeyDeleteActionExpr -> TableConstraint

data ForeignKeyActionExpr

restrictExpr :: ForeignKeyActionExpr

cascadeExpr :: ForeignKeyActionExpr

setNullExpr :: ForeignKeyActionExpr

setDefaultExpr :: ForeignKeyActionExpr

data ForeignKeyDeleteActionExpr

foreignKeyDeleteActionExpr :: ForeignKeyActionExpr -> ForeignKeyDeleteActionExpr

data ForeignKeyUpdateActionExpr

foreignKeyUpdateActionExpr :: ForeignKeyActionExpr -> ForeignKeyUpdateActionExpr

data CreateSequenceExpr

createSequenceExpr :: Qualified SequenceName -> Maybe IncrementByExpr -> Maybe MinValueExpr -> Maybe MaxValueExpr -> Maybe StartWithExpr -> Maybe CacheExpr -> Maybe CycleExpr -> CreateSequenceExpr

data AlterSequenceExpr

alterSequenceExpr :: Qualified SequenceName -> Maybe IncrementByExpr -> Maybe MinValueExpr -> Maybe MaxValueExpr -> Maybe StartWithExpr -> Maybe CacheExpr -> Maybe CycleExpr -> AlterSequenceExpr

data IncrementByExpr

incrementBy :: Int64 -> IncrementByExpr

data MinValueExpr

minValue :: Int64 -> MinValueExpr

noMinValue :: MinValueExpr

data MaxValueExpr

maxValue :: Int64 -> MaxValueExpr

noMaxValue :: MaxValueExpr

data StartWithExpr

startWith :: Int64 -> StartWithExpr

data CacheExpr

cache :: Int64 -> CacheExpr

data CycleExpr

cycle :: CycleExpr

noCycle :: CycleExpr

cycleIfTrue :: Bool -> CycleExpr

data DropSequenceExpr

dropSequenceExpr :: Maybe IfExists -> Qualified SequenceName -> DropSequenceExpr

nextVal :: Qualified SequenceName -> ValueExpression

nextValFunction :: FunctionName

currVal :: Qualified SequenceName -> ValueExpression

currValFunction :: FunctionName

setVal :: Qualified SequenceName -> Int64 -> ValueExpression

setValFunction :: FunctionName

data ReturningExpr

returningExpr :: SelectList -> ReturningExpr

data TableExpr

tableExpr :: Qualified TableName -> Maybe WhereClause -> Maybe OrderByClause -> Maybe GroupByClause -> Maybe LimitExpr -> Maybe OffsetExpr -> TableExpr

data SelectList

selectColumns :: [ColumnName] -> SelectList

data DerivedColumn

deriveColumn :: ValueExpression -> DerivedColumn

deriveColumnAs :: ValueExpression -> ColumnName -> DerivedColumn

selectDerivedColumns :: [DerivedColumn] -> SelectList

selectStar :: SelectList

data QueryExpr

queryExpr :: SelectClause -> SelectList -> Maybe TableExpr -> QueryExpr

data SelectClause

selectClause :: SelectExpr -> SelectClause

data SelectExpr

selectExpr :: Maybe Distinct -> SelectExpr

data Distinct

Constructors

Distinct

data OrderByExpr

appendOrderByExpr :: OrderByExpr -> OrderByExpr -> OrderByExpr

orderByExpr :: RawSql -> OrderByDirection -> OrderByExpr

orderByColumnsExpr :: NonEmpty (ColumnName, OrderByDirection) -> OrderByExpr

data OrderByDirection

data NullsOrder

Constructors

NullsFirst

NullsLast

ascendingOrder :: OrderByDirection

descendingOrder :: OrderByDirection

ascendingOrderWith :: NullsOrder -> OrderByDirection

descendingOrderWith :: NullsOrder -> OrderByDirection

data OrderByClause

orderByClause :: OrderByExpr -> OrderByClause

data OffsetExpr

offsetExpr :: Int -> OffsetExpr

data TableName

tableName :: String -> TableName

data SequenceName

sequenceName :: String -> SequenceName

data SchemaName

schemaName :: String -> SchemaName

data SavepointName

savepointName :: String -> SavepointName

data Qualified name

qualified :: IdentifierExpression name => Maybe SchemaName -> name -> Qualified name

data IndexName

indexName :: String -> IndexName

data Identifier

identifier :: String -> Identifier

identifierFromBytes :: ByteString -> Identifier

unquotedIdentifier :: String -> Identifier

unquotedIdentifierFromBytes :: ByteString -> Identifier

class IdentifierExpression name where

Methods

toIdentifier :: name -> Identifier

fromIdentifier :: Identifier -> name

data FunctionName

functionName :: String -> FunctionName

data CursorName

cursorName :: String -> CursorName

data ConstraintName

constraintName :: String -> ConstraintName

data ColumnName

columnName :: String -> ColumnName

data LimitExpr

limitExpr :: Int -> LimitExpr

data InsertExpr

insertExpr :: Qualified TableName -> Maybe InsertColumnList -> InsertSource -> Maybe ReturningExpr -> InsertExpr

data InsertColumnList

insertColumnList :: [ColumnName] -> InsertColumnList

data InsertSource

insertSqlValues :: [[SqlValue]] -> InsertSource

data CreateIndexExpr

createIndexExpr :: IndexUniqueness -> Qualified TableName -> NonEmpty ColumnName -> CreateIndexExpr

data IndexUniqueness

Constructors

UniqueIndex

NonUniqueIndex

data DropIndexExpr

dropIndexExpr :: IndexName -> DropIndexExpr

createNamedIndexExpr :: IndexUniqueness -> Qualified TableName -> IndexName -> RawSql -> CreateIndexExpr

data IfExists

ifExists :: IfExists

data GroupByExpr

appendGroupByExpr :: GroupByExpr -> GroupByExpr -> GroupByExpr

groupByExpr :: RawSql -> GroupByExpr

groupByColumnsExpr :: NonEmpty ColumnName -> GroupByExpr

data GroupByClause

groupByClause :: GroupByExpr -> GroupByClause

data DeleteExpr

deleteExpr :: Qualified TableName -> Maybe WhereClause -> Maybe ReturningExpr -> DeleteExpr

data DeclareExpr

declare :: CursorName -> Maybe ScrollExpr -> Maybe HoldExpr -> QueryExpr -> DeclareExpr

data ScrollExpr

scroll :: ScrollExpr

noScroll :: ScrollExpr

data HoldExpr

withHold :: HoldExpr

withoutHold :: HoldExpr

data CloseExpr

close :: Either AllCursors CursorName -> CloseExpr

data AllCursors

allCursors :: AllCursors

data FetchExpr

fetch :: Maybe CursorDirection -> CursorName -> FetchExpr

data MoveExpr

move :: Maybe CursorDirection -> CursorName -> MoveExpr

data CursorDirection

next :: CursorDirection

prior :: CursorDirection

first :: CursorDirection

last :: CursorDirection

absolute :: Int -> CursorDirection

relative :: Int -> CursorDirection

count :: Int -> CursorDirection

fetchAll :: CursorDirection

forward :: CursorDirection

forwardCount :: Int -> CursorDirection

forwardAll :: CursorDirection

backward :: CursorDirection

backwardCount :: Int -> CursorDirection

backwardAll :: CursorDirection

data ColumnDefinition

columnDefinition :: ColumnName -> DataType -> Maybe ColumnConstraint -> Maybe ColumnDefault -> ColumnDefinition

data ColumnConstraint

notNullConstraint :: ColumnConstraint

nullConstraint :: ColumnConstraint

data ColumnDefault

columnDefault :: ValueExpression -> ColumnDefault

data DataType

timestampWithZone :: DataType

timestampWithoutZone :: DataType

date :: DataType

tsvector :: DataType

varchar :: Int32 -> DataType

char :: Int32 -> DataType

text :: DataType

uuid :: DataType

boolean :: DataType

doublePrecision :: DataType

bigSerial :: DataType

bigInt :: DataType

serial :: DataType

int :: DataType

smallint :: DataType

oid :: DataType