Orville.PostgreSQL.Internal.FieldDefinition

Description

Documentation

data FieldDefinition nullability a

FieldDefinition determines the SQL constsruction of a column in the database, comprising the name, SQL type and whether the field is nullable. A FieldDefinition is matched to a particular Haskell type, which it knows how to marshall to and from the database representation of SQL type for the field.

fieldName :: FieldDefinition nullability a -> FieldName

The name used in database queries to reference the field.

fieldDescription :: FieldDefinition nullability a -> Maybe String

Returns the description that was passed to setFieldDescription, if any.

setFieldDescription :: String -> FieldDefinition nullability a -> FieldDefinition nullability a

Sets the description for the field. This description not currently used anywhere by Orville itself, but users can retrieve the description via fieldDescription for their own purposes (e.g. generating documentation).

fieldType :: FieldDefinition nullability a -> SqlType a

The SqlType for the FieldDefinition determines the PostgreSQL data type used to define the field as well as how to mashall Haskell values to and from the database.

fieldIsNotNullable :: FieldDefinition nullability a -> Bool

Indicates whether a field is nullable.

fieldDefaultValue :: FieldDefinition nullability a -> Maybe (DefaultValue a)

Returns the default value definition for the field, if any has been set.

fieldNullability :: FieldDefinition nullability a -> FieldNullability a

Resolves the nullablity of a field to a concrete type, which is returned via the FieldNullability type. You can pattern match on this type to then extract the either Nullable or NotNull not field for cases where you may require different logic based on the nullability of a field.

setField :: FieldDefinition nullability a -> a -> SetClause

Constructs a SetClause that will set the column named in the field definition to the given value. The value is be converted to SQL value using fieldValueToSqlValue

(.:=) :: FieldDefinition nullability a -> a -> SetClause

Operator alias for setField

data FieldNullability a

A FieldNullability is returned by the fieldNullability function, which can be used when a function works on both Nullable and NotNull functions but needs to deal with each type of field separately. It adds wrapper constructors around the FieldDefinition that you can pattern match on to then work with a concrete Nullable or NotNull field.

Constructors

NullableField (FieldDefinition Nullable a)

NotNullField (FieldDefinition NotNull a)

fieldValueToSqlValue :: FieldDefinition nullability a -> a -> SqlValue

Mashalls a Haskell value to be stored in the field to its SqlValue representation.

fieldValueFromSqlValue :: FieldDefinition nullability a -> SqlValue -> Either String a

Marshalls a SqlValue from the database into the Haskell value that represents it. This may fail, in which case a Left is returned with an error message.

fieldColumnName :: FieldDefinition nullability a -> ColumnName

Constructs the ColumnName for a field for use in SQL expressions from the Expr module.

fieldColumnDefinition :: FieldDefinition nullability a -> ColumnDefinition

Constructions the equivalant FieldDefinition as a SQL expression, generally for use in DDL for creating column in a table.

data FieldName

stringToFieldName :: String -> FieldName

fieldNameToString :: FieldName -> String

fieldNameToColumnName :: FieldName -> ColumnName

fieldNameToByteString :: FieldName -> ByteString

byteStringToFieldName :: ByteString -> FieldName

data NotNull

NotNull is a values-less type used to track that a FieldDefinition represents a field that is marked not-null in the database schema. See the Nullability type for the value-level representation of field nullability.

data Nullable

Nullable is a values-less type used to track that a FieldDefinition represents a field that is marked nullable in the database schema. See the Nullability type for the value-level representation of field nullability.

convertField :: (SqlType a -> SqlType b) -> FieldDefinition nullability a -> FieldDefinition nullability b

Applies a SqlType conversion to a FieldDefinition. You can use this function the create FieldDefinitions for based on the primitive ones provided, but with more specific Haskell types.

See convertSqlType and tryConvertSqlType for functions to create the conversion needed as the first argument to convertField.

coerceField :: (Coercible a b, Coercible b a) => FieldDefinition nullability a -> FieldDefinition nullability b

A specialization of convertField that can be used with types that implement Coercible. This is particularly useful for newtype wrappers around primitive types.

nullableField :: FieldDefinition NotNull a -> FieldDefinition Nullable (Maybe a)

Makes a NotNull field Nullable by wrapping the Haskell type of the field in Maybe. The field will be marked as NULL in the database schema and the value Nothing will be used to represent NULL values when converting to and from sql.

asymmetricNullableField :: FieldDefinition Nullable a -> FieldDefinition Nullable (Maybe a)

Adds a Maybe wrapper to a field that is already nullable. (If your field is NotNull, you wanted nullableField instead of this function). Note that fields created using this function have asymetric encoding and decoding of NULL values. Because the provided field is Nullable, NULL values decode from the database already have a representation in the a type, so NULL will be decoded as 'Just of type a for NULL'. This means if you insert a Nothing value using the field, it will be read back as Just value. This is useful for building high level combinators that might need to make fields Nullable but need the value to be decoded in its underlying type when reading back (e.g. maybeMapper from SqlMarshaller).

setDefaultValue :: DefaultValue a -> FieldDefinition nullability a -> FieldDefinition nullability a

Sets a default value for the field. The default value will be added as part of the column definition in the database. Because the default value is ultimately provided by the database this can be used to add a not-null column to safely to an existing table as long as a reasonable default value is available to use.

removeDefaultValue :: FieldDefinition nullability a -> FieldDefinition nullability a

Removes any default value that may have been set on a field via setDefaultValue.

prefixField :: String -> FieldDefinition nullability a -> FieldDefinition nullability a

Adds a prefix, followed by an underscore, to a field's name.

integerField

Arguments

:: String
Name of the field in the database

-> FieldDefinition NotNull Int32

Builds a FieldDefinition that stores Haskell Int32 values as the PostgreSQL INT type.

serialField

Arguments

:: String
Name of the field in the database

-> FieldDefinition NotNull Int32

Builds a FieldDefinition that stores an Int32 value as the SERIAL type. This can be used to create auto-incrementing columns.

smallIntegerField

Arguments

:: String
Name of the field in the database

-> FieldDefinition NotNull Int16

Builds a FieldDefinition that stores Haskell Int16 values as the PostgreSQL SMALLINT type.

bigIntegerField

Arguments

:: String
Name of the field in the database

-> FieldDefinition NotNull Int64

Builds a FieldDefinition that stores Haskell Int64 values as the PostgreSQL BIGINT type.

bigSerialField

Arguments

:: String
Name of the field in the database

-> FieldDefinition NotNull Int64

Builds a FieldDefinition that stores an Int64 value as the BIGSERIAL type. This can be used to create auto-incrementing columns.

doubleField

Arguments

:: String
Name of the field in the database

-> FieldDefinition NotNull Double

Builds a FieldDefinition that stores a Double value as the "DOUBLE PRECISION" type. Note: PostgreSQL's "DOUBLE PRECISION" type only allows for up to 15 digits of precision, so some rounding may occur when values are stored in the database.

booleanField

Arguments

:: String
Name of the field in the database

-> FieldDefinition NotNull Bool

Builds a FieldDefinition that stores Haskell Bool values as the PostgreSQL BOOLEAN type.

unboundedTextField

Arguments

:: String
Name of the field in the database

-> FieldDefinition NotNull Text

Builds a FieldDefinition that stores Haskell Text values as the PostgreSQL TEXT type. Note that this PostgreSQL has no particular limit on the length of text stored.

boundedTextField

Arguments

:: String
Name of the field in the database

-> Int32
Maximum length of text in the field

-> FieldDefinition NotNull Text

Builds a FieldDefinition that stores Haskell Text values as the PostgreSQL VARCHAR type. Attempting to store a value beyond the length specified will cause an error.

- TODO: We should have a test for this.

fixedTextField

Arguments

:: String
Name of the field in the database

-> Int32
Maximum length of text in the field

-> FieldDefinition NotNull Text

Builds a FieldDefinition that stores Haskell Text values as the PostgreSQL CHAR type. Attempting to store a value beyond the length specified will cause an error. Storing a value that is not the full length of the field will result in padding by the database.

- TODO: We should have a test for this.

textSearchVectorField :: String -> FieldDefinition NotNull Text

TODO: write meaningful docs for this when we build a better Haskell definition for representing text search vectors.

dateField

Arguments

:: String
Name of the field in the database

-> FieldDefinition NotNull Day

Builds a FieldDefinition that stores Haskell Day values as the PostgreSQL DATE type.

utcTimestampField

Arguments

:: String
Name of the field in the database

-> FieldDefinition NotNull UTCTime

Builds a FieldDefinition that stores Haskell 'Time.UTCTime values as the PostgreSQL "TIMESTAMP with time zone" type.

localTimestampField

Arguments

:: String
Name of the field in the database

-> FieldDefinition NotNull LocalTime

Builds a FieldDefinition that stores Haskell 'Time.UTCTime values as the PostgreSQL "TIMESTAMP without time zone" type.

uuidField

Arguments

:: String
Name of the field in the database

-> FieldDefinition NotNull UUID

Builds a FieldDefinition that stores Haskell UUID values as the PostgreSQL UUID type.

fieldOfType

Arguments

:: SqlType a
SqlType that represents the PostgreSQL data type for the field.

-> String
Name of the field in the database

-> FieldDefinition NotNull a

Builds a FieldDefinition for will use the given SqlType to determine the database representation of the field. If you have created a custom SqlType, you can use this function to construct a helper like the other functions in this module for creating FieldDefinitions for your custom type.