Orville.PostgreSQL.Internal.OrvilleState

Documentation

data OrvilleState

OrvilleState is used to manange opening connections to the database, transactions, etc. newOrvilleState should be used to create an appopriate initial state for your monad's context.

newOrvilleState :: ErrorDetailLevel -> Pool Connection -> OrvilleState

Creates a appropriate initial OrvilleState that will use the connection pool given to initiate connections to the database.

resetOrvilleState :: OrvilleState -> OrvilleState

Creates a new initial OrvilleState using the connection pool from the provide state. You might need to use this if you are spawning one Orville monad from another and they should not share the same connection and transaction state.

orvilleConnectionPool :: OrvilleState -> Pool Connection

orvilleConnectionState :: OrvilleState -> ConnectionState

orvilleErrorDetailLevel :: OrvilleState -> ErrorDetailLevel

orvilleTransactionCallback :: OrvilleState -> TransactionEvent -> IO ()

orvilleSqlCommenterAttributes :: OrvilleState -> Maybe SqlCommenterAttributes

addTransactionCallback :: (TransactionEvent -> IO ()) -> OrvilleState -> OrvilleState

Registers a callback to be invoked during transactions.

The callback given will be called after the SQL statement corresponding to the given event has finished executing. Callbacks will be called in the order the are added.

Note: There is no specialized error handling for these callbacks. This means that if a callback raises an exception no further callbacks will be called and the exception will propagate up until it caught elsewhere. In particular, if an exception is raised by a callback upon opening the transaction it will cause the transaction to be rolled-back the same as any other exception that might happen during the transaction. In general, we recommend only using callbacks that either raise no exceptions or can handle their own exceptions cleanly.

data TransactionEvent

Describes an event in the lifecycle of a database transaction. You can use addTransactionCallBack to register a callback to respond to these events. The callback will be called after the even in question has been succesfully executed.

Constructors

BeginTransaction
Indicates a new transaction has been started

NewSavepoint Savepoint
Indicates that a new savepoint has been saved within a transaction

ReleaseSavepoint Savepoint
Indicates that a previous savepoint has been released. It can no longer be rolled back to.

RollbackToSavepoint Savepoint
Indicates that rollbac was performed to a prior savepoint.

Note: It is possible to rollback to a savepoint prior to the most recent one without releasing or rolling back to intermediate savepoints. Doing so destroys any savepoints created after given savepoint. Although Orville currently always matches NewSavepoint with either ReleaseSavepoint or RollbackToSavepoint, it is recommended that you do not rely on this behavior.

CommitTransaction
Indicates that the transaction has been committed.

RollbackTransaction
Indicates that the transaction has been rolled back.

openTransactionEvent :: TransactionState -> TransactionEvent

rollbackTransactionEvent :: TransactionState -> TransactionEvent

transactionSuccessEvent :: TransactionState -> TransactionEvent

class HasOrvilleState m where

HasOrvilleState is the typeclass that Orville uses to access and manange the connection pool and state tracking when it is being executed inside an unknown Monad. It is a specialized version of the Reader interface so that it can easily implemented by application Monads that already have a Reader context and want to simply add OrvilleState as an attribute to that context, like so

   data MyApplicationState =
     MyApplicationState
       { appConfig :: MyAppConfig
       , appOrvilleState :: OrvilleState
       }

   newtype MyApplicationMonad a =
     MyApplicationMonad (ReaderT MyApplicationState IO) a

   instance HasOrvilleState MyApplicationMonad where
     askOrvilleState =
       MyApplicationMonad (asks appOrvilleState)

     localOrvilleState f (MyApplicationMonad reader) =
       MyApplicationMonad $
         local
           (state -> state { appOrvilleState = f (appOrvilleState state))
           reader
 

An instance for 'ReaderT OrvilleState m' is provided as a convenience in the case that your application has no extra context to track.

Methods

askOrvilleState :: m OrvilleState

localOrvilleState

Arguments

:: (OrvilleState -> OrvilleState)
The function to modify the OrvilleState

-> m a
The monad operation to execute with the modified state

-> m a

data ConnectionState

INTERNAL: This type is used to signal whether a database connection has been retrieved from the pool for the current operation or not. The value is tracked in the OrvilleState for the host monad, and is checked by withConnection to avoid checking out two separate connections for a multiple operations that needs to be run on the same connection (e.g. multiple operations inside a transaction).

Constructors

NotConnected

Connected ConnectedState

data ConnectedState

Constructors

ConnectedState

Fields

connectedConnection :: Connection

connectedTransaction :: Maybe TransactionState

connectState :: ConnectedState -> OrvilleState -> OrvilleState

INTERNAL: Transitions the OrvilleState into "connected" status, storing the given Connection as the database connection to be used to execute all queries. This is used by withConnection to track the connection it retrieves from the pool.

data TransactionState

Constructors

OutermostTransaction

SavepointTransaction Savepoint

newTransaction :: Maybe TransactionState -> TransactionState

data Savepoint

A internal Orville identifier for a savepoint in a PostgreSQL transaction.

savepointNestingLevel :: Savepoint -> Int

initialSavepoint :: Savepoint

nextSavepoint :: Savepoint -> Savepoint

orvilleSqlExecutionCallback :: OrvilleState -> forall a. QueryType -> RawSql -> IO a -> IO a

addSqlExecutionCallback :: (forall a. QueryType -> RawSql -> IO a -> IO a) -> OrvilleState -> OrvilleState

Adds a callback to be called when an Orville operation executes a SQL statement. The callback is given the IO action that will perform the query execution and must call that action for the query to be run. In particular, you can use this to time query and log any that are slow.

Calls to any previously added callbacks will also be execute as part of the IO action passed to the new callback. Thus the newly added callback happens "around" the previously added callback.

There is no special exception handling done for these callbacks beyond what they implement themelves. Any callbacks should allow for the possibility that the IO action they are given may raise an exception.

orvilleBeginTransactionExpr :: OrvilleState -> BeginTransactionExpr

setBeginTransactionExpr :: BeginTransactionExpr -> OrvilleState -> OrvilleState

Sets the SQL expression that Orville will use to begin transactions. You can control the transaction isolation level by building your own BeginTransactionExpr with the desired isolation level.

setSqlCommenterAttributes :: SqlCommenterAttributes -> OrvilleState -> OrvilleState

Sets the SqlCommenterAttributes that Orville will then add to any following statement executions.

Since: 0.10.0

addSqlCommenterAttributes :: SqlCommenterAttributes -> OrvilleState -> OrvilleState

Adds the SqlCommenterAttributes to the already existing that Orville will then add to any following statement executions.

Since: 0.10.0