Orville.PostgreSQL.Internal.PrimaryKey

Documentation

data PrimaryKey key

A Haskell description of the FieldDefinitions that make up the primary key of a SQL table. This type supports composite primary keys as well as singular ones.

primaryKeyDescription :: PrimaryKey key -> String

primaryKeyDescription builds a user-readable representation of the primary key for use in error messages and such. It is a comma-delimited list of the names of the fields that make up the primary key.

primaryKeyFieldNames :: PrimaryKey key -> NonEmpty FieldName

Retrieves the names of the fields that are part of the primary key.

primaryKeyToSql :: PrimaryKey key -> key -> NonEmpty SqlValue

primaryKeyToSql converts a Haskell value for a primary key into the (possibly multiple) sql values that represent the primary key in the database.

primaryKey :: FieldDefinition NotNull key -> PrimaryKey key

primaryKey constructs a single-field primary key from the FieldDefinition that corresponds to the primary key's column. This is generally used while building a TableDefinition.

data PrimaryKeyPart key

A PrimaryKeyPart describes one field of a composite primary key. Values are built using primaryKeyPart and then used with compositePrimaryKey to build a PrimaryKey

compositePrimaryKey :: PrimaryKeyPart key -> [PrimaryKeyPart key] -> PrimaryKey key

compositePrimaryKey constructs a multi-field primary key from the given parts, each of which corresponds to one field in the primary key. You should use this while building a TableDefinition for a table that you want to have a multi-column primary key. See primaryKeyPart for how to build the parts to be passed as parameters. Note: there is no special significance to the first argument other than requiring that there is at least one field in the primary key.

primaryKeyPart :: (key -> part) -> FieldDefinition NotNull part -> PrimaryKeyPart key

primaryKeyPart constructs a building block for a composite primary key based a FieldDefinition and an accessor function to extract the value for that field from the Haskell key type that represents the overall composite key. PrimaryKeyPart values built using this function are usually then passed in a list to compositePrimaryKey to build a PrimaryKey.

mapPrimaryKeyParts :: (forall part. (key -> part) -> FieldDefinition NotNull part -> a) -> PrimaryKey key -> NonEmpty a

mapPrimaryKeyParts provides a way to access the innards of a PrimaryKey definition to extract information. The given function will be called on each part of the primary key in order and the list of results is returned. Note that single-field and multi-field primary keys are treated the same by this function, with the single-field case simply behaving as composite key with just one part.

mkPrimaryKeyExpr :: PrimaryKey key -> PrimaryKeyExpr

Builds a PrimaryKeyExpr that is suitable to be used when creating a table to define the primary key on the table.

primaryKeyEquals :: PrimaryKey key -> key -> WhereCondition

primaryKeyEquals builds a WhereCondition that will match the row where the primary key is equal to the given value. For single-field primary keys this is equivalent to fieldEquals, but primaryKeyEquals also handles composite primary keys.

primaryKeyEqualsExpr :: PrimaryKey key -> key -> BooleanExpr

Like primaryKeyEquals, but returns the condition as a lower-level BooleanExpr