Orville.PostgreSQL.Internal.RawSql

Description

The funtions in this module are named with the intent that it is imported qualified as RawSql.

Documentation

data RawSql

RawSql provides a type for efficiently constructing raw sql statements from smaller parts and then executing them. It also supports using placeholder values to pass parameters with a query without having to interpolate them as part of the actual sql state and being exposed to sql injection.

parameter :: SqlValue -> RawSql

Includes an input parameter in the RawSql statement that will be passed using placeholders (e.g. '$1') rather than being included directly in the sql statement. This is the correct way to include input from untrusted sources as part of a RawSql query. The parameter must be formatted in a textual representation, which the database will interpret. The database type for the value will be inferred by the database based on its usage in the query.

fromString :: String -> RawSql

Constructs a RawSql from a String value using utf8 encoding.

Note that because the string is treated as raw sql it completely up to the caller to protected againt sql-injections attacks when using this function. Never use this function with input read from an untrusted source.

fromText :: Text -> RawSql

Constructs a RawSql from a Text value using utf8 encoding.

Note that because the text is treated as raw sql it completely up to the caller to protected againt sql-injections attacks when using this function. Never use this function with input read from an untrusted source.

fromBytes :: ByteString -> RawSql

Constructs a RawSql from a ByteString value, which is assumed to be encoded sensibly for the database to handle.

Note that because the string is treated as raw sql it completely up to the caller to protected againt sql-injections attacks when using this function. Never use this function with input read from an untrusted source.

intercalate :: (SqlExpression sql, Foldable f) => RawSql -> f sql -> RawSql

Concatenates a list of RawSql values using another RawSql value as the a separator between the items.

execute :: SqlExpression sql => Connection -> sql -> IO Result

Executes a RawSql value using the executeRaw function. Make sure to read the documentation of executeRaw for caveats and warnings. Use with caution.

Note that because this is done in IO no callback functions are available to be called.

executeVoid :: SqlExpression sql => Connection -> sql -> IO ()

Executes a RawSql value using the executeRawVoid function. Make sure to read the documentation of executeRawVoid for caveats and warnings. Use with caution.

Note that because this is done in IO no callback functions are available to be called.

connectionEscaping :: Connection -> Escaping IO

Escaping done in IO based using the escaping functions provided by the connection, which can apply escaping based on the specific connection properties.

If you don't have a connection available and are only planning on using the SQL for explanatory or example purposes, see exampleEscaping.

Fragments provided for convenience

space :: RawSql

Just a plain old space, provided for convenience

comma :: RawSql

Just a plain old comma, provided for convenience

commaSpace :: RawSql

Comma space separator, provided for convenience

leftParen :: RawSql

Just a plain old left paren, provided for convenience

rightParen :: RawSql

Just a plain old right paren, provided for convenience

dot :: RawSql

Just a plain period, provided for convenience

doubleQuote :: RawSql

Just a plain double quote, provided for convenience

doubleColon :: RawSql

Just two colons, provided for convenience

stringLiteral :: ByteString -> RawSql

Includes a bytestring value as string literal in the SQL statement. The string literal will be escaped and quoted for you, the value provided does not need to include surrounding quotes or escape special characters.

Note: It's better to use the parameter function where possible to pass values to be used as input to a SQL statement. There are some situations where PostgreSQL does not allow this, however (for instance, in some DDL statements). This function is provided for those situations.

Integer values as literals

intDecLiteral :: Int -> RawSql

Constructs a RawSql from an Int value. The integral value is included directly in the SQL string, not passed as a parameter. When dealing with user input it is better to use parameter rather whenever possible.

int8DecLiteral :: Int8 -> RawSql

Constructs a RawSql from an Int8 value. The integral value is included directly in the SQL string, not passed as a parameter. When dealing with user input it is better to use parameter rather whenever possible.

int16DecLiteral :: Int16 -> RawSql

Constructs a RawSql from an Int16 value. The integral value is included directly in the SQL string, not passed as a parameter. When dealing with user input it is better to use parameter rather whenever possible.

int32DecLiteral :: Int32 -> RawSql

Constructs a RawSql from an Int32 value. The integral value is included directly in the SQL string, not passed as a parameter. When dealing with user input it is better to use parameter rather whenever possible.

int64DecLiteral :: Int64 -> RawSql

Constructs a RawSql from an Int64 value. The integral value is included directly in the SQL string, not passed as a parameter. When dealing with user input it is better to use parameter rather whenever possible.

Generic interface for generating sql

class SqlExpression a where

Methods

toRawSql :: a -> RawSql

unsafeFromRawSql :: RawSql -> a

toBytesAndParams :: (SqlExpression sql, Monad m) => Escaping m -> sql -> m (ByteString, [Maybe PgTextFormatValue])

Constructs the actual SQL bytestring and parameter values that will be passed to the database to execute a RawSql query. Any string literals thar are included in the SQL expression will be escaping using the given escaping directive.

toExampleBytes :: SqlExpression sql => sql -> ByteString

Builds the bytes that represent the raw sql. These bytes may not be executable on their own, because they may contain placeholders that must be filled in, but can be useful for inspecting sql queries.

data Escaping m

Provides procedures for escaping parts of a raw SQL query so that they can be safely executed. Escaping may be done in some Monad m, allowing for the use of escaping operations provided by Connection, which operate in the IO monad.

See connectionEscaping and exampleEscaping.

Constructors

Escaping

Fields

escapeStringLiteral :: ByteString -> m ByteString

exampleEscaping :: Escaping Identity

Escaping done in pure Haskell that is suitable for showing SQL examples, but is not guaranteed to be sufficient for all database connections. For escaping that is based on the actual connection to the database, see connectionEscaping.