Orville.PostgreSQL.Internal.SelectOptions

Documentation

data WhereCondition

A WhereCondition represents a SQL expression that can be added to SELECT statements to restrict the rows returned by a query. Various functions are provided below to construct conditions.

fieldEquals :: FieldDefinition nullability a -> a -> WhereCondition

Checks that the value in a field equals a particular value.

(.==) :: FieldDefinition nullability a -> a -> WhereCondition infixl 9

Operator alias for fieldEquals

fieldNotEquals :: FieldDefinition nullability a -> a -> WhereCondition

Checks that the value in a field does not equal a particular value.

(./=) :: FieldDefinition nullability a -> a -> WhereCondition infixl 9

Operator alias for fieldNotEquals

fieldGreaterThan :: FieldDefinition nullability a -> a -> WhereCondition

Checks that the value in a field is greater than a particular value.

(.>) :: FieldDefinition nullability a -> a -> WhereCondition infixl 9

Operator alias for fieldGreaterThan

fieldLessThan :: FieldDefinition nullability a -> a -> WhereCondition

Checks that the value in a field is less than a particular value.

(.<) :: FieldDefinition nullability a -> a -> WhereCondition infixl 9

Operator alias for fieldLessThan

fieldGreaterThanOrEqualTo :: FieldDefinition nullability a -> a -> WhereCondition

Checks that the value in a field is greater than or equal to a particular value.

(.>=) :: FieldDefinition nullability a -> a -> WhereCondition infixl 9

Operator alias for fieldGreaterThanOrEqualTo

fieldLessThanOrEqualTo :: FieldDefinition nullability a -> a -> WhereCondition

Checks that the value in a field is less than or equal to a particular value.

(.<=) :: FieldDefinition nullability a -> a -> WhereCondition infixl 9

Operator alias for fieldLessThanOrEqualTo

fieldIsNull :: FieldDefinition Nullable a -> WhereCondition

Checks that the value in a field is null.

fieldIsNotNull :: FieldDefinition Nullable a -> WhereCondition

Checks that the value in a field is not null.

fieldLike :: FieldDefinition nullability a -> Text -> WhereCondition

Checks that the value in a field matches a like pattern

fieldLikeInsensitive :: FieldDefinition nullability a -> Text -> WhereCondition

Checks that the value in a field matches a like pattern case insensitively

fieldIn :: FieldDefinition nullability a -> NonEmpty a -> WhereCondition

Checks that a field matches a list of values

(.<-) :: FieldDefinition nullability a -> NonEmpty a -> WhereCondition infixl 9

Operator alias for fieldIn

fieldNotIn :: FieldDefinition nullability a -> NonEmpty a -> WhereCondition

Checks that a field does not match a list of values

(.</-) :: FieldDefinition nullability a -> NonEmpty a -> WhereCondition infixl 9

Operator alias for fieldNotIn

fieldTupleIn :: FieldDefinition nullabilityA a -> FieldDefinition nullabilityB b -> NonEmpty (a, b) -> WhereCondition

Checks that a tuple of two fields is in the list of specified tuplies

fieldTupleNotIn :: FieldDefinition nullabilityA a -> FieldDefinition nullabilityB b -> NonEmpty (a, b) -> WhereCondition

Checks that a tuple of two fields is not in the list of specified tuplies

whereAnd :: WhereCondition -> WhereCondition -> WhereCondition

Combines multiple WhereConditions together using AND.

(.&&) :: WhereCondition -> WhereCondition -> WhereCondition infixr 8

Operator alias for whereAnd

whereOr :: WhereCondition -> WhereCondition -> WhereCondition

Combines multiple WhereConditions together using 'OR.

(.||) :: WhereCondition -> WhereCondition -> WhereCondition infixr 8

Operator alias for 'whereOr

whereBooleanExpr :: BooleanExpr -> WhereCondition

Constructs a WhereCondition from a BooleanExpr. You can use this to together with the RawSql related functions in the Expr module to use SQL expressions that Orville does not directly support in your queries.

whereConditionToBooleanExpr :: WhereCondition -> BooleanExpr

Converts a WhereCondition into a BooleanExpr so that you can use it with the functions from the Expr module if you are building SQL expressions yourself.

data SelectOptions

A SelectOptions is a set of options that can be used to change the way a basic query function works by adding WHERE, 'ORDER BY', 'GROUP BY', etc. Functions are provided to construct SelectOptions for individual options, which may then be combined via <> (also exposed as appendSelectOptions).

emptySelectOptions :: SelectOptions

A set of empty SelectOptions that will not change how a query is run.

appendSelectOptions :: SelectOptions -> SelectOptions -> SelectOptions

Combines multple select options together, unioning the options together where possible. For options where this is not possible, (e.g. LIMIT), the one on the left is preferred.

selectDistinct :: SelectOptions -> SelectClause

Builds the SelectClause that should be used to include the distincts from the SelectOptions on a query.

selectWhereClause :: SelectOptions -> Maybe WhereClause

Builds the WhereClause that should be used to include the WhereConditions from the SelectOptions on a query. This will be Nothing where no WhereConditions have been specified.

selectOrderByClause :: SelectOptions -> Maybe OrderByClause

Builds the OrderByClause that should be used to include the OrderByClauses from the SelectOptions on a query. This will be Nothing where no OrderByClauses have been specified.

selectGroupByClause :: SelectOptions -> Maybe GroupByClause

Builds the GroupByClause that should be used to include the GroupByClauses from the SelectOptions on a query. This will be Nothing where no GroupByClauses have been specified.

selectLimitExpr :: SelectOptions -> Maybe LimitExpr

Builds a LimitExpr that will limit the query results to the number specified in the SelectOptions (if any)

selectOffsetExpr :: SelectOptions -> Maybe OffsetExpr

Builds a OffsetExpr that will limit the query results to the number specified in the SelectOptions (if any)

distinct :: SelectOptions

Constructs a SelectOptions with just distinct set to True.

where_ :: WhereCondition -> SelectOptions

Constructs a SelectOptions with just the given WhereCondition.

orderBy :: OrderBy -> SelectOptions

Constructs a SelectOptions with just the given OrderBy.

limit :: Int -> SelectOptions

Constructs a SelectOptions that will apply the given limit

offset :: Int -> SelectOptions

Constructs a SelectOptions that will apply the given offset

groupBy :: GroupByExpr -> SelectOptions

Constructs a SelectOptions with just the given GroupByClause.

data OrderBy

data OrderByDirection

data NullsOrder

Constructors

NullsFirst

NullsLast

ascendingOrder :: OrderByDirection

ascendingOrderWith :: NullsOrder -> OrderByDirection

descendingOrder :: OrderByDirection

descendingOrderWith :: NullsOrder -> OrderByDirection

orderByOrderByExpr :: OrderByExpr -> OrderBy

orderByField :: FieldDefinition nullability value -> OrderByDirection -> OrderBy

orderByColumnName :: ColumnName -> OrderByDirection -> OrderBy

appendOrderBy :: OrderBy -> OrderBy -> OrderBy

orderByToClause :: OrderBy -> OrderByClause

orderByToExpr :: OrderBy -> OrderByExpr